requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
fair-esm>=2.0.0
transformers>=4.30.0

# Structural Biology
biopython>=1.81

# LLM Integration
openai>=1.0.0
anthropic>=0.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# API & Web
requests>=2.31.0
aiohttp>=3.8.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Parallel Processing
joblib>=1.3.0

# Optional: Graph Neural Networks (for future Phase 3)
# torch-geometric>=2.3.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
